Artifacts and stuff
Holy shit, making it so that artifacts with index numbers greater than 255 work was an unholy pain in the butt!!!

+++ src/xtra2.c
@@ -4229,7 +4229,7 @@ void monster_death(int m_idx)
 		}
 		else
 		{
-			byte a_idx = 0;
+			int a_idx = 0;
 			int chance = 0;
 			int I_kind = 0;
